Refactor input fields into seperate files 4 better organization.
